+++ CHANGELOG.md
@@ -13,6 +13,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - plugin state API
 - ability to edit custom actions
 - ability to use `await` operator in custom actions
+- hiding of panel title from component settings
+
+### Changed
+
+- the design of the panel
+
+### Removed
+
+- unnecessary component BtnBack
 
 ## [0.5.0]